+++ src/main/java/org/leisureup/global/auth/social/internal/GoogleApiErrorDecoder.java
@@ -1,6 +1,8 @@
 package org.leisureup.global.auth.social.internal;
 
-public class GoogleApiErrorDecoder extends DefaultOAuthApiErrorDecoder {
+import org.leisureup.global.response.external.*;
+
+public class GoogleApiErrorDecoder extends DefaultFeignErrorDecoder {
 
     private static final String MSG_PREFIX = """
             Google 로그인 중 문제가 발생했습니다.
