+++ src/main/java/org/leisureup/global/auth/social/internal/KakaoApiErrorDecoder.java
@@ -1,6 +1,8 @@
 package org.leisureup.global.auth.social.internal;
 
-public class KakaoApiErrorDecoder extends DefaultOAuthApiErrorDecoder {
+import org.leisureup.global.response.external.*;
+
+public class KakaoApiErrorDecoder extends DefaultFeignErrorDecoder {
 
     private static final String MSG_PREFIX = """
             Kakao 로그인 중 문제가 발생했습니다.
